FR-1016 : As a Case Worker I want to Amend Case Details (#137)
