+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/controllers/UpdateContestedCaseController.java
@@ -13,9 +13,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
 import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
-import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseDocument;
-import uk.gov.hmcts.reform.finrem.caseorchestration.service.ConsentOrderService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.OnlineFormDocumentService;
 
 import java.util.ArrayList;
@@ -25,12 +23,11 @@ import java.util.Map;
 import static java.util.Objects.nonNull;
 import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CCDConfigConstant.MINI_FORM_A;
-import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CCDConfigConstant.STATE;
 
 @RestController
 @RequestMapping(value = "/case-orchestration")
 @Slf4j
-public class CaseMaintenanceController implements BaseController {
+public class UpdateContestedCaseController implements BaseController {
 
     private static final String DIVORCE_STAGE_REACHED = "divorceStageReached";
     private static final String DIVORCE_UPLOAD_EVIDENCE_2 = "divorceUploadEvidence2";
@@ -42,51 +39,19 @@ public class CaseMaintenanceController implements BaseController {
     @Autowired
     private OnlineFormDocumentService onlineFormDocumentService;
 
-    @Autowired
-    private ConsentOrderService consentOrderService;
-
-    @PostMapping(path = "/update-case", consumes = APPLICATION_JSON, produces = APPLICATION_JSON)
-    @ApiOperation(value = "Handles update Case details and cleans up the data fields based on the options choosen")
-    @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "Callback was processed successFully or in case of an error message is "
-                    + "attached to the case",
-                    response = AboutToStartOrSubmitCallbackResponse.class),
-            @ApiResponse(code = 400, message = "Bad Request"),
-            @ApiResponse(code = 500, message = "Internal Server Error")})
-    public ResponseEntity<AboutToStartOrSubmitCallbackResponse> updateCase(
-            @RequestHeader(value = "Authorization", required = false) String authToken,
-            @RequestBody CallbackRequest ccdRequest) {
-
-        log.info("Received request for updateCase ");
-        validateCaseData(ccdRequest);
-        Map<String, Object> caseData = ccdRequest.getCaseDetails().getData();
-        updateDivorceDetails(caseData);
-        updatePeriodicPaymentData(caseData);
-        updatePropertyDetails(caseData);
-        updateRespondentSolicitorAddress(caseData);
-        updateD81Details(caseData);
-        updateLatestConsentOrder(ccdRequest);
-        return ResponseEntity.ok(AboutToStartOrSubmitCallbackResponse.builder().data(caseData).build());
-    }
-
-    private void updateLatestConsentOrder(CallbackRequest callbackRequest) {
-        CaseDetails caseDetails = callbackRequest.getCaseDetails();
-        Map<String, Object> caseData = caseDetails.getData();
-        caseData.put("latestConsentOrder", consentOrderService.getLatestConsentOrderData(callbackRequest));
-    }
 
     @PostMapping(path = "/update-contested-case", consumes = APPLICATION_JSON, produces = APPLICATION_JSON)
     @ApiOperation(value = "Handles update Contested Case details and cleans up the data fields"
-            + " based on the options choosen")
+        + " based on the options choosen")
     @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "Callback was processed successFully or in case of an error message is "
-                    + "attached to the case",
-                    response = AboutToStartOrSubmitCallbackResponse.class),
-            @ApiResponse(code = 400, message = "Bad Request"),
-            @ApiResponse(code = 500, message = "Internal Server Error")})
+        @ApiResponse(code = 200, message = "Callback was processed successFully or in case of an error message is "
+            + "attached to the case",
+            response = AboutToStartOrSubmitCallbackResponse.class),
+        @ApiResponse(code = 400, message = "Bad Request"),
+        @ApiResponse(code = 500, message = "Internal Server Error")})
     public ResponseEntity<AboutToStartOrSubmitCallbackResponse> updateContestedCase(
-            @RequestHeader(value = "Authorization", required = false) String authToken,
-            @RequestBody CallbackRequest ccdRequest) {
+        @RequestHeader(value = "Authorization", required = false) String authToken,
+        @RequestBody CallbackRequest ccdRequest) {
 
         log.info("Received request for contested - updateCase ");
         validateCaseData(ccdRequest);
@@ -101,7 +66,7 @@ public class CaseMaintenanceController implements BaseController {
         updateContestedMiamDetails(caseData);
         cleanupAdditionalDocuments(caseData);
         CaseDocument document = onlineFormDocumentService.generateDraftContestedMiniFormA(authToken,
-                ccdRequest.getCaseDetails());
+            ccdRequest.getCaseDetails());
         caseData.put(MINI_FORM_A, document);
         return ResponseEntity.ok(AboutToStartOrSubmitCallbackResponse.builder().data(caseData).build());
     }
@@ -197,16 +162,16 @@ public class CaseMaintenanceController implements BaseController {
     private void updateMiamExceptionDetails(Map<String, Object> caseData) {
         ArrayList miamExemptionsChecklist = (ArrayList) caseData.get("MIAMExemptionsChecklist");
         removeExemptionCheckLists(caseData, miamExemptionsChecklist,
-                "other", "MIAMOtherGroundsChecklist");
+            "other", "MIAMOtherGroundsChecklist");
 
         removeExemptionCheckLists(caseData, miamExemptionsChecklist,
-                "domesticViolence", "MIAMDomesticViolenceChecklist");
+            "domesticViolence", "MIAMDomesticViolenceChecklist");
 
         removeExemptionCheckLists(caseData, miamExemptionsChecklist,
-                "urgency", "MIAMUrgencyReasonChecklist");
+            "urgency", "MIAMUrgencyReasonChecklist");
 
         removeExemptionCheckLists(caseData, miamExemptionsChecklist,
-                "previousMIAMattendance", "MIAMPreviousAttendanceChecklist");
+            "previousMIAMattendance", "MIAMPreviousAttendanceChecklist");
     }
 
     private void removeExemptionCheckLists(Map<String, Object> caseData, ArrayList miamExemptionsChecklist,
@@ -275,18 +240,6 @@ public class CaseMaintenanceController implements BaseController {
     }
 
 
-    private void updateDivorceDetails(Map<String, Object> caseData) {
-        if (caseData.get(DIVORCE_STAGE_REACHED).equals("Decree Nisi")) {
-            // remove Decree Absolute details
-            caseData.put(DIVORCE_UPLOAD_EVIDENCE_2, null);
-            caseData.put(DIVORCE_DECREE_ABSOLUTE_DATE, null);
-        } else {
-            // remove Decree Nisi details
-            caseData.put(DIVORCE_UPLOAD_EVIDENCE_1, null);
-            caseData.put(DIVORCE_DECREE_NISI_DATE, null);
-        }
-    }
-
     private void updateDivorceDetailsForContestedCase(Map<String, Object> caseData) {
         if (equalsTo((String) caseData.get(DIVORCE_STAGE_REACHED), "Decree Nisi")) {
             // remove Decree Absolute details
@@ -308,49 +261,6 @@ public class CaseMaintenanceController implements BaseController {
         }
     }
 
-    private void updatePeriodicPaymentData(Map<String, Object> caseData) {
-        List natureOfApplication2 = (List) caseData.get("natureOfApplication2");
-
-        if (hasNotSelected(natureOfApplication2, "Periodical Payment Order")) {
-            removePeriodicPaymentData(caseData);
-        } else {
-            // if written agreement for order for children
-            if (equalsTo((String) caseData.get("natureOfApplication5"), "Yes")) {
-                caseData.put("natureOfApplication6", null);
-                caseData.put("natureOfApplication7", null);
-            }
-        }
-    }
-
-    private void updatePropertyDetails(Map<String, Object> caseData) {
-        List natureOfApplication2 = (List) caseData.get("natureOfApplication2");
-
-        if (hasNotSelected(natureOfApplication2, "Property Adjustment  Order")) {
-            removePropertyAdjustmentDetails(caseData);
-        }
-    }
-
-    private void updateD81Details(Map<String, Object> caseData) {
-        if (equalsTo((String) caseData.get("d81Question"), "Yes")) {
-            caseData.put("d81Applicant", null);
-            caseData.put("d81Respondent", null);
-        } else {
-            caseData.put("d81Joint", null);
-        }
-    }
-
-    private void updateRespondentSolicitorAddress(Map<String, Object> caseData) {
-        if (equalsTo((String) caseData.get("appRespondentRep"), "No")) {
-            removeRespondentSolicitorAddress(caseData);
-        }
-    }
-
-    private void removePeriodicPaymentData(Map<String, Object> caseData) {
-        caseData.put("natureOfApplication5", null);
-        caseData.put("natureOfApplication6", null);
-        caseData.put("natureOfApplication7", null);
-        caseData.put("orderForChildrenQuestion1", null);
-    }
 
     private void removeRespondentSolicitorAddress(Map<String, Object> caseData) {
         caseData.put("rSolicitorName", null);
@@ -362,10 +272,6 @@ public class CaseMaintenanceController implements BaseController {
         caseData.put("rSolicitorDXnumber", null);
     }
 
-    private void removePropertyAdjustmentDetails(Map<String, Object> caseData) {
-        caseData.put("natureOfApplication3a", null);
-        caseData.put("natureOfApplication3b", null);
-    }
 
     private void updateContestedRespondentDetails(Map<String, Object> caseData) {
         if (equalsTo((String) caseData.get("respondentRepresented"), "No")) {
